chore: Move PreferenceBasedCriteria and ExampleBasedCriteria to their own files
Related to https://github.com/shaka-project/shaka-player/issues/7768
